requirements.txt
certifi==2018.4.16
chardet==3.0.4
daemon==1.2
idna==2.6
influxdb==5.2.0
kafka-python==1.4.2
requests==2.18.4
redis==3.0.1
urllib3==1.22
pylint==2.4.2
graypy==2.1.0
sphinx==2.2.0
sphinx-rtd-theme==0.4.3
